api/requirements.txt
Flask==3.1.0
python-dotenv==1.1.0
bcrypt==4.2.0
Jinja2==3.1.6
Werkzeug==3.1.3
SQLAlchemy==2.0.20
Flask-Limiter==3.12
cryptography==44.0.2
supabase==2.28.0
